Bugfix: unbound env variables when running finalize hook
Inside snippets buffers, env variables like
`ptemplate-source-directory' and `ptemplate-target-directory' are set
up by `ptemplate--setup-snippet-env', but outside snippets buffers, we
can't refer to these variables, e.g. in finalize-hooks, we use
`ptemplate-target' to get the absolute path of :open files but
`ptemplate-target-directory' is unbounded (actually is nil) thus
causing `ptemplate-target' to return a relative path.

In order to fix this, setup the env variables inside a temp buffer
and run the finalize hook inside the temp buffer.